转成了float模式，发现当batch增大时与普通cnn时间差别较大，空间差不多，应该是batch的并行问题

+++ conv2d/conv2d_modules/conv2d_model.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 from torch.nn.modules.module import Module
 from torch.nn.modules.utils import _pair
 
-from conv2d_double.conv2d_functions import Conv2dFunction
+from conv2d.conv2d_functions import Conv2dFunction
 
 
 class Conv2d(Module):
@@ -32,11 +32,11 @@ class Conv2d(Module):
         if out_channels % groups != 0:
             raise ValueError('out_channels must be divisible by groups')
 
-        self.weight = nn.Parameter(torch.cuda.DoubleTensor(out_channels, in_channels // groups, *self.kernel_size))
+        self.weight = nn.Parameter(torch.cuda.FloatTensor(out_channels, in_channels // groups, *self.kernel_size))
         nn.init.kaiming_normal(self.weight.data, mode='fan_out')
         nn.init.constant(self.weight.data, 1)
         if bias:
-            self.bias = nn.Parameter(torch.cuda.DoubleTensor(out_channels))
+            self.bias = nn.Parameter(torch.cuda.FloatTensor(out_channels))
             nn.init.uniform(self.bias.data, -0.1, 0.1)
         else:
             self.register_parameter('bias', None)
