+++ conv2d/lib/conv2d_cuda_kernel.h
@@ -9,7 +9,7 @@
 //#include "TH/TH.h"
 
 void im2col(cudaStream_t stream,
-           const double *data_in,
+           const float *data_in,
            const int input_c,
            const int input_h, const int input_w,
            const int output_h, const int output_w,
@@ -17,11 +17,11 @@ void im2col(cudaStream_t stream,
            const int pad_h, const int pad_w,
            const int stride_h, const int stride_w,
            const int dilation_h, const int dilation_w,
-           double *data_col);
+           float *data_col);
 
 
 void col2im(cudaStream_t stream,
-         const double *data_col,
+         const float *data_col,
          const int input_c,
          const int input_h, const int input_w,
          const int output_h, const int output_w,
@@ -29,5 +29,5 @@ void col2im(cudaStream_t stream,
          const int pad_h, const int pad_w,
          const int stride_h, const int stride_w,
          const int dilation_h, const int dilation_w,
-         double *grad_im);
+         float *grad_im);
 //#endif //DEFORM3D_NEW_PYTORCH_DEFORM_CONV3D_CUDA_BACKWARD_CU_H
